Convert replication calls from action.execute to nodeclient
Replication relies on some internal actions to do local replication
operations. This commit converts these actions to registered transport
actions so NodeClient can be used.
